* 'feature/FIJI-2219' of git.ringcentral.com:Fiji/Fiji:
  feat(FIJI-1637): [Upload file from computer] should only clean item ids in the post.
  Merge branch 'feature/FIJI-1637' into feature/FIJI-1985
  feat(FIJI-1637): [Upload file from computer] add CANCELED status and remove item id when it was canceled
  fix(FIJI-2313): [Should show notification' uploaded a version N 'and the previous files should be updated to the new version when updating the file] fix FIJI-2313
  fix(FIJI-2313): [Should show notification' uploaded a version N 'and the previous files should be updated to the new version when updating the file] fix always show update file issue
  fix(FIJI-2313): [Should show notification' uploaded a version N 'and the previous files should be updated to the new version when updating the file] no need to notify UI when post updated to UI.
  fix(FIJI-2313): [Should show notification' uploaded a version N 'and the previous files should be updated to the new version when updating the file] cancel upload task when the task is updated
  fix(FIJI-2313): [Should show notification' uploaded a version N 'and the previous files should be updated to the new version when updating the file] only check exist item with post ids.
  fix(FIJI-2313): [Should show notification' uploaded a version N 'and the previous files should be updated to the new version when updating the file] should start send item to server when user send post.
  fix(FIJI-2313): [Should show notification' uploaded a version N 'and the previous files should be updated to the new version when updating the file] fix UT issues
  fix(FIJI-2313): [Should show notification' uploaded a version N 'and the previous files should be updated to the new version when updating the file] fix error in file version
  fix(FIJI-2313): [Should show notification' uploaded a version N 'and the previous files should be updated to the new version when updating the file] fix file version issue and remove status from Item

+++ packages/sdk/src/service/item/itemFileUploadHandler.ts
@@ -2,6 +2,7 @@
  * @Author: Jerry Cai (jerry.cai@ringcentral.com)
  * @Date: 2018-12-05 09:30:00
  */
+import _ from 'lodash';
 import { NETWORK_FAIL_TYPE, mainLogger } from 'foundation';
 import { StoredFile, ItemFile, Item, Raw, Progress } from '../../models';
 import AccountService from '../account';
@@ -32,43 +33,89 @@ class ItemFileUploadHandler {
     isUpdate: boolean,
   ): Promise<ItemFile | null> {
     if (file.has(FILE_FORM_DATA_KEYS.FILE)) {
-      const itemFile = this._toItemFile(
-        groupId,
-        file,
-        isUpdate,
-        SENDING_STATUS.INPROGRESS,
-      );
-      await this._preSaveItemFile(itemFile);
-      this._sendItemFile(groupId, itemFile, file, isUpdate);
+      const itemFile = this._toItemFile(groupId, file, isUpdate);
+      await this._preSaveItemFile(itemFile, file);
+      this._sendItemFile(itemFile, file);
       return itemFile;
     }
     return null;
   }
 
+  private _getCachedItem(itemId: number) {
+    const cache = this._progressCaches.get(itemId);
+    return cache ? cache.itemFile : undefined;
+  }
+
+  async sendItemData(groupId: number, postItemIds: number[]) {
+    const needWaitItemIds: number[] = [];
+    postItemIds.forEach((id: number) => {
+      const itemStatus = this._progressCaches.get(id);
+      if (itemStatus && itemStatus.itemFile) {
+        const item = itemStatus.itemFile;
+        if (this._hasValidStoredFile(item)) {
+          this._uploadItem(groupId, item, this._isUpdateItem(item));
+        } else {
+          needWaitItemIds.push(item.id);
+        }
+      }
+    });
+
+    if (needWaitItemIds.length > 0) {
+      this._waitUntilAllItemCreated(groupId, needWaitItemIds);
+    }
+  }
+
+  private _waitUntilAllItemCreated(
+    groupId: number,
+    uploadingItemFileIds: number[],
+  ) {
+    const listener = (params: {
+      success: boolean;
+      preInsertId: number;
+      updatedId: number;
+    }) => {
+      // handle item upload result.
+      const { success, preInsertId, updatedId } = params;
+      if (
+        !uploadingItemFileIds.includes(preInsertId) ||
+        updatedId !== preInsertId
+      ) {
+        return;
+      }
+      _.remove(uploadingItemFileIds, (id: number) => {
+        return id === preInsertId;
+      });
+
+      const item = this._getCachedItem(preInsertId);
+      if (success && item) {
+        this._uploadItem(groupId, item, this._isUpdateItem(item));
+      }
+
+      if (uploadingItemFileIds.length === 0) {
+        notificationCenter.removeListener(
+          SERVICE.ITEM_SERVICE.PSEUDO_ITEM_STATUS,
+          listener,
+        );
+      }
+    };
+
+    notificationCenter.on(SERVICE.ITEM_SERVICE.PSEUDO_ITEM_STATUS, listener);
+  }
   async resendFailedFile(itemId: number) {
+    this._updateFileProgress(itemId, SENDING_STATUS.INPROGRESS);
+
     const itemDao = daoManager.getDao(ItemDao);
     const itemInDB = (await itemDao.get(itemId)) as ItemFile;
     let sendFailed = false;
     if (itemInDB) {
-      await this._partialUpdateItemFile({
-        id: itemInDB.id,
-        _id: itemInDB.id,
-        status: SENDING_STATUS.INPROGRESS,
-      });
-
       const groupId = itemInDB.group_ids[0];
       const isUpdate = itemInDB.is_new;
-      const version = itemInDB.versions[0];
-      if (
-        version &&
-        version.download_url.length > 0 &&
-        version.url.length > 0
-      ) {
+      if (this._hasValidStoredFile(itemInDB)) {
         await this._uploadItem(groupId, itemInDB, isUpdate);
       } else {
         const cacheItem = this._progressCaches.get(itemId);
         if (groupId && cacheItem && cacheItem.file) {
-          await this._sendItemFile(groupId, itemInDB, cacheItem.file, isUpdate);
+          await this._sendItemFile(itemInDB, cacheItem.file);
         } else {
           sendFailed = true;
         }
@@ -102,6 +149,8 @@ class ItemFileUploadHandler {
       }
     });
 
+    this._emitItemFileStatus(SENDING_STATUS.CANCELED, itemId, itemId);
+
     const itemDao = daoManager.getDao(ItemDao);
     await itemDao.delete(itemId);
     notificationCenter.emitEntityDelete(ENTITY.ITEM, [itemId]);
@@ -116,8 +165,23 @@ class ItemFileUploadHandler {
     return status ? status.progress : undefined;
   }
 
-  cleanUploadingFiles(groupId: number) {
-    this._uploadingFiles.delete(groupId);
+  cleanUploadingFiles(groupId: number, toRemoveItemIds: number[]) {
+    let itemFiles = this._uploadingFiles.get(groupId);
+    if (itemFiles) {
+      itemFiles = _.differenceWith(
+        itemFiles,
+        toRemoveItemIds,
+        (item: ItemFile, id: number) => {
+          return id === item.id;
+        },
+      );
+
+      if (itemFiles.length === 0) {
+        this._uploadingFiles.delete(groupId);
+      } else {
+        this._uploadingFiles.set(groupId, itemFiles);
+      }
+    }
   }
 
   getItemsSendStatus(itemIds: number[]): SENDING_STATUS[] {
@@ -141,10 +205,41 @@ class ItemFileUploadHandler {
     return result;
   }
 
-  private _updateFailedFIleProgress(failedItemId: number) {
+  public async getUpdateItemVersion(itemFile: ItemFile) {
+    let versionNumber = 0;
+    if (itemFile) {
+      if (itemFile.id > 0) {
+        versionNumber = itemFile.versions.length;
+      } else {
+        const existItemFile = await this._getOldestExistFile(
+          itemFile.group_ids[0],
+          itemFile.name,
+        );
+        versionNumber = existItemFile ? existItemFile.versions.length + 1 : 0;
+      }
+    }
+
+    return versionNumber;
+  }
+
+  private _updateFileProgress(failedItemId: number, status: SENDING_STATUS) {
     const info = this._progressCaches.get(failedItemId);
     if (info && info.progress) {
-      info.progress.loaded = -1;
+      let loaded = info.progress.loaded;
+      switch (status) {
+        case SENDING_STATUS.FAIL:
+          loaded = -1;
+          break;
+        case SENDING_STATUS.INPROGRESS:
+          loaded = 0;
+          break;
+        case SENDING_STATUS.SUCCESS:
+          loaded = info.progress.total;
+          break;
+        default:
+          break;
+      }
+      info.progress.loaded = loaded;
     }
   }
 
@@ -199,7 +294,6 @@ class ItemFileUploadHandler {
   private async _uploadFileToAmazonS3(
     formFile: FormData,
     preInsertItem: ItemFile,
-    isUpdate: boolean,
     requestHolder: RequestHolder,
   ) {
     const groupId = preInsertItem.group_ids[0];
@@ -221,10 +315,10 @@ class ItemFileUploadHandler {
         requestHolder,
       );
       if (uploadResponse.isOk()) {
-        this._onUploadFileSuccess(
+        this._handleFileUploadSuccess(
           extendFileData.stored_file,
+          groupId,
           preInsertItem,
-          isUpdate,
         );
       } else {
         this._handleItemFileSendFailed(itemId, uploadResponse);
@@ -237,7 +331,6 @@ class ItemFileUploadHandler {
   private async _uploadFileFileToGlip(
     file: FormData,
     preInsertItem: ItemFile,
-    isUpdate: boolean,
     requestHolder: RequestHolder,
   ) {
     const groupId = preInsertItem.group_ids[0];
@@ -251,10 +344,10 @@ class ItemFileUploadHandler {
     );
 
     if (uploadRes.isOk()) {
-      await this._onUploadFileSuccess(
+      await this._handleFileUploadSuccess(
         uploadRes.unwrap()[0],
+        groupId,
         preInsertItem,
-        isUpdate,
       );
     } else {
       this._handleItemFileSendFailed(preInsertItem.id, uploadRes);
@@ -262,57 +355,20 @@ class ItemFileUploadHandler {
     }
   }
 
-  private async _onUploadFileSuccess(
-    storedFile: StoredFile,
-    preInsertItem: ItemFile,
-    isUpdate: boolean,
-  ) {
-    const groupId = preInsertItem.group_ids[0];
-    await this._handleFileUploadSuccess(storedFile, groupId, preInsertItem);
-    await this._uploadItem(groupId, preInsertItem, isUpdate);
+  private _getRequestHolder(preInsertItemId: number) {
+    const uploadInfo = this._progressCaches.get(
+      preInsertItemId,
+    ) as ItemFileUploadStatus;
+    return uploadInfo.requestHolder;
   }
 
-  private async _sendItemFile(
-    groupId: number,
-    preInsertItem: ItemFile,
-    file: FormData,
-    isUpdate: boolean,
-  ) {
-    const requestHolder: RequestHolder = { request: undefined };
-    const progress = {
-      id: preInsertItem.id,
-      total: 0,
-      loaded: 0,
-    };
-
-    const preInsertItemId = preInsertItem.id;
-    const status = this._progressCaches.get(preInsertItemId);
-    if (status) {
-      status.requestHolder = requestHolder;
-      status.progress = progress;
-      status.file = file;
-    } else {
-      this._progressCaches.set(preInsertItemId, {
-        requestHolder,
-        progress,
-        file,
-      });
-    }
+  private async _sendItemFile(preInsertItem: ItemFile, file: FormData) {
+    const requestHolder = this._getRequestHolder(preInsertItem.id);
 
     if (isInBeta(EBETA_FLAG.BETA_S3_DIRECT_UPLOADS)) {
-      await this._uploadFileToAmazonS3(
-        file,
-        preInsertItem,
-        isUpdate,
-        requestHolder,
-      );
+      await this._uploadFileToAmazonS3(file, preInsertItem, requestHolder);
     } else {
-      await this._uploadFileFileToGlip(
-        file,
-        preInsertItem,
-        isUpdate,
-        requestHolder,
-      );
+      await this._uploadFileFileToGlip(file, preInsertItem, requestHolder);
     }
   }
 
@@ -357,13 +413,23 @@ class ItemFileUploadHandler {
     const fileVersion = this._toFileVersion(storedFile);
     preInsertItem.versions = [fileVersion];
     this._updateUploadingFiles(groupId, preInsertItem);
+    this._updateCachedFilesStatus(preInsertItem);
 
     itemDao.update(preInsertItem);
+    const itemId = preInsertItem.id;
     await this._partialUpdateItemFile({
-      id: preInsertItem.id,
-      _id: preInsertItem.id,
+      id: itemId,
+      _id: itemId,
       versions: [fileVersion],
     });
+    this._emitItemFileStatus(SENDING_STATUS.INPROGRESS, itemId, itemId);
+  }
+
+  private _updateCachedFilesStatus(newItemFile: ItemFile) {
+    const status = this._progressCaches.get(newItemFile.id);
+    if (status) {
+      status.itemFile = newItemFile;
+    }
   }
 
   private _updateUploadingFiles(groupId: number, newItemFile: ItemFile) {
@@ -407,7 +473,7 @@ class ItemFileUploadHandler {
     const replaceItemFiles = new Map<number, ItemFile>();
     replaceItemFiles.set(preInsertId, itemFile);
     notificationCenter.emitEntityReplace(ENTITY.ITEM, replaceItemFiles);
-    this._emitItemFileStatus(true, preInsertId, itemFile.id);
+    this._emitItemFileStatus(SENDING_STATUS.SUCCESS, preInsertId, itemFile.id);
   }
 
   private _handleItemFileSendFailed<T>(
@@ -419,15 +485,9 @@ class ItemFileUploadHandler {
       return;
     }
 
-    this._updateFailedFIleProgress(preInsertId);
+    this._updateFileProgress(preInsertId, SENDING_STATUS.FAIL);
 
-    this._emitItemFileStatus(false, preInsertId, preInsertId);
-
-    this._partialUpdateItemFile({
-      id: preInsertId,
-      _id: preInsertId,
-      status: SENDING_STATUS.FAIL,
-    });
+    this._emitItemFileStatus(SENDING_STATUS.FAIL, preInsertId, preInsertId);
   }
 
   private _partialUpdateItemFile(updateData: object) {
@@ -452,15 +512,66 @@ class ItemFileUploadHandler {
     };
   }
 
-  private async _preSaveItemFile(newItemFile: ItemFile) {
-    const groupId = newItemFile.group_ids[0];
+  private _isUpdateItem(itemFile: ItemFile) {
+    return !itemFile.is_new;
+  }
+  private _saveItemFileToUploadingFiles(preInsertItem: ItemFile) {
+    const groupId = preInsertItem.group_ids[0];
     let existFiles: ItemFile[] | undefined = this._uploadingFiles.get(groupId);
     if (existFiles && existFiles.length > 0) {
-      existFiles.push(newItemFile);
+      if (this._isUpdateItem(preInsertItem)) {
+        const toCancelIds: number[] = [];
+        existFiles = existFiles.filter((item: ItemFile) => {
+          if (item.name === preInsertItem.name) {
+            toCancelIds.push(item.id);
+            return false;
+          }
+          return true;
+        });
+
+        toCancelIds.forEach((id: number) => {
+          this.cancelUpload(id);
+        });
+      }
+      existFiles.push(preInsertItem);
     } else {
-      existFiles = [newItemFile];
+      existFiles = [preInsertItem];
     }
     this._uploadingFiles.set(groupId, existFiles);
+  }
+
+  private _saveItemFileToProgressCache(
+    preInsertItem: ItemFile,
+    formFile: FormData,
+  ) {
+    const requestHolder: RequestHolder = { request: undefined };
+    const progress = {
+      id: preInsertItem.id,
+      total: 0,
+      loaded: 0,
+    };
+
+    const preInsertItemId = preInsertItem.id;
+    const status = this._progressCaches.get(preInsertItemId);
+    if (status) {
+      status.requestHolder = requestHolder;
+      status.progress = progress;
+      status.file = formFile;
+      status.itemFile = preInsertItem;
+    } else {
+      this._progressCaches.set(preInsertItemId, {
+        requestHolder,
+        progress,
+        file: formFile,
+        itemFile: preInsertItem,
+      });
+    }
+  }
+
+  private async _preSaveItemFile(newItemFile: ItemFile, formFile: FormData) {
+    this._saveItemFileToUploadingFiles(newItemFile);
+    this._saveItemFileToProgressCache(newItemFile, formFile);
+
     const itemDao = daoManager.getDao(ItemDao);
     await itemDao.put(newItemFile);
   }
@@ -468,8 +579,7 @@ class ItemFileUploadHandler {
   private _toItemFile(
     groupId: number,
     formFile: FormData,
-    isNew: boolean,
-    sendStatus: SENDING_STATUS,
+    isUpdate: boolean,
   ): ItemFile {
     const file = formFile.get(FILE_FORM_DATA_KEYS.FILE) as File;
     const accountService: AccountService = AccountService.getInstance();
@@ -479,11 +589,10 @@ class ItemFileUploadHandler {
     const id = GlipTypeUtil.generatePseudoIdByType(TypeDictionary.TYPE_ID_FILE);
     return {
       id,
-      sendStatus,
       created_at: now,
       modified_at: now,
       creator_id: userId,
-      is_new: isNew,
+      is_new: !isUpdate,
       deactivated: false,
       version: versionHash(),
       group_ids: [groupId],
@@ -517,12 +626,12 @@ class ItemFileUploadHandler {
   }
 
   private _emitItemFileStatus(
-    success: boolean,
+    status: SENDING_STATUS,
     preInsertId: number,
     updatedId: number,
   ) {
     notificationCenter.emit(SERVICE.ITEM_SERVICE.PSEUDO_ITEM_STATUS, {
-      success,
+      status,
       preInsertId,
       updatedId,
     });
@@ -530,7 +639,7 @@ class ItemFileUploadHandler {
 
   private async _updateItem(existItem: ItemFile, preInsertItem: ItemFile) {
     existItem.is_new = false;
-    existItem.versions = existItem.versions.concat(preInsertItem.versions);
+    existItem.versions = preInsertItem.versions.concat(existItem.versions);
     existItem.modified_at = Date.now();
     existItem._id = existItem.id;
     delete existItem.id;
@@ -555,6 +664,11 @@ class ItemFileUploadHandler {
       return sorted[0];
     }
     return null;
+  }
+
+  private _hasValidStoredFile(itemFile: ItemFile) {
+    const version = itemFile.versions[0];
+    return version && version.download_url.length > 0 && version.url.length > 0;
   }
 }
